app/src/test/java/com/example/myopenweather/data/repository/DefaultWeatherRepositoryTest.kt
package com.example.myopenweather.data.repository

import com.example.myopenweather.data.mappers.toWeather
import com.example.myopenweather.data.network.WeatherApi

import junit.framework.TestCase.assertEquals

import com.example.myopenweather.data.sampleForecastResponse
import com.example.myopenweather.domain.model.Weather
import com.example.myopenweather.domain.utils.Result

import kotlinx.coroutines.test.runTest
import org.junit.Before
import org.junit.Test
import io.mockk.coEvery
import io.mockk.mockk

class DefaultWeatherRepositoryTest {
    private lateinit var repository: DefaultWeatherRepository
    private val weatherApi = mockk<WeatherApi>()

    @Before
    fun setup() {
        repository = DefaultWeatherRepository(weatherApi)
    }

    @Test
    fun `when getWeatherForecast is called, it should emit loading state and then success state`() =
        runTest {
            coEvery {
                weatherApi.getWeatherForecast(
                    any(),
                    any(),
                    any()
                )
            } returns sampleForecastResponse

            val results = mutableListOf<Result<Weather>>()
            repository.getWeatherForecast("Munich").collect { result ->
                results.add(result)
            }
            assertEquals(Result.Loading, results[0])

            assertEquals(Result.Success(sampleForecastResponse.toWeather()), results[1])
        }
}